- Refactoring as discussed in PR - Fixing tests
