+++ src/ladds/distributedMemParallelization/DomainDecomposition.cpp
@@ -5,7 +5,9 @@
  */
 #include "DomainDecomposition.h"
 
+#include <autopas/utils/WrapOpenMP.h>
 #include <autopas/utils/inBox.h>
+#include <spdlog/spdlog.h>
 
 std::array<double, 3> LADDS::DomainDecomposition::getGlobalBoxMin() const {
   return globalBoxMin;
@@ -27,4 +29,20 @@ bool LADDS::DomainDecomposition::containsLocal(const std::array<double, 3> &coor
 }
 autopas::AutoPas_MPI_Comm LADDS::DomainDecomposition::getCommunicator() const {
   return communicator;
+}
+
+void LADDS::DomainDecomposition::printMPIInfo() const {
+  int rank{};
+  autopas::AutoPas_MPI_Comm_rank(this->getCommunicator(), &rank);
+  if (rank == 0) {
+    auto logger = spdlog::get(LADDS_SPD_LOGGER_NAME);
+    int numRanks{};
+    autopas::AutoPas_MPI_Comm_size(AUTOPAS_MPI_COMM_WORLD, &numRanks);
+    SPDLOG_LOGGER_INFO(logger.get(),
+                       "Parallelization Configuration\n"
+                       "MPI Ranks              : {}\n"
+                       "OpenMP Threads per Rank: {}\n",
+                       numRanks,
+                       autopas::autopas_get_max_threads());
+  }
 }
