+++ src/ladds/distributedMemParallelization/RegularGridDecomposition.h
@@ -30,5 +30,18 @@ class RegularGridDecomposition : public DomainDecomposition {
    * @return tuple{dimensions, periods, grid coordinates of current rank}
    */
   std::tuple<std::array<int, 3>, std::array<int, 3>, std::array<int, 3>> getGridInfo() const;
+
+  /**
+   * Send the given list of leaving particles to the respective target ranks and receive their leaving
+   * particles which are relevant for the local rank.
+   * @param leavingParticles in/out parameter of leaving particles. If everything worked the vector should be empty
+   * after the function call.
+   * @param autopas
+   * @param decomposition
+   * @return Vector of incoming particles.
+   */
+  std::vector<LADDS::Particle> communicateParticles(std::vector<LADDS::Particle> &leavingParticles,
+                                                    autopas::AutoPas<Particle> &autopas,
+                                                    const DomainDecomposition &decomposition) const override;
 };
 }  // namespace LADDS
